feat: Add CustomerServiceUsage resource and related logic (master)
Introduces a new Filament resource for managing customer service usage.
This includes:

- Creating `ManageCustomerServiceUsage` page.
- Adding a `customerServiceUsages` relationship to `CustomerService`.
- Updating the `CustomerServiceUsage` model cast for `days_of_usage`.
- Modifying the migration for `customer_service_usages` table to use integer for `days_of_usage`.
- Renaming `InvoiceHistory` page to `ManageInvoiceHistory`.
- Adjusting navigation items in `CustomerServiceResource`.
- Adding comments to `MakeRecurringInvoiceCommand` for clarity on service filtering.

+++ database/migrations/2025_12_21_212302_create_customer_service_usages_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
             $table->unsignedBigInteger('invoice_id');
             $table->dateTime('used_since');
             $table->dateTime('next_billing_date');
-            $table->decimal('days_of_usage', 8, 0);
+            $table->integer('days_of_usage');
             $table->decimal('daily_price', 20, 0);
             $table->decimal('total_price', 20, 0);
             $table->timestamps();
